+++ engines/wintermute/metaengine.cpp
@@ -131,24 +131,8 @@ public:
 		return retVal;
 	}
 
-	const Common::AchievementsInfo getAchievementsInfo(const Common::String &target) const override {
-		Common::String gameId = ConfMan.get("gameid", target);
-
-		Common::AchievementsPlatform platform = Common::STEAM_ACHIEVEMENTS;
-		if (ConfMan.get("extra", target).contains("GOG")) {
-			platform = Common::GALAXY_ACHIEVEMENTS;
-		}
-
-		// "(gameId, platform) -> result" search
-		Common::AchievementsInfo result;
-		for (const AchievementDescriptionList *i = achievementDescriptionList; i->gameId; i++) {
-			if (i->gameId == gameId && i->platform == platform) {
-				result.platform = i->platform;
-				result.appId = i->appId;
-				break;
-			}
-		}
-		return result;
+	const Common::AchievementDescriptionList* getAchievementDescriptionList() const override {
+		return Wintermute::achievementDescriptionList;
 	}
 
 	Common::KeymapArray initKeymaps(const char *target) const override {
